+++ pkg/authentication/certificate.go
@@ -68,6 +68,20 @@ func NewCertificate(publicKeyPem, privateKeyPem string) (Certificate, error) {
 	}, nil
 }
 
+// creates a certificate from only the public key, in the case of the aspsp public cert to validate signatures
+func NewPublicCertificate(publicKeyPem string) (Certificate, error) {
+	publicKey, err := jwt.ParseRSAPublicKeyFromPEM([]byte(publicKeyPem))
+	if err != nil {
+		return nil, errors.Wrap(err, "error with public key")
+	}
+	publicPem := []byte(publicKeyPem)
+
+	return &certificate{
+		publicKey:     publicKey,
+		publicCertPem: publicPem,
+	}, nil
+}
+
 func (c certificate) PublicKey() *rsa.PublicKey {
 	return c.publicKey
 }
@@ -123,6 +137,9 @@ func (c certificate) SignatureIssuer(tpp bool) (string, error) {
 	}
 
 	if tpp {
+		if ou == "" {
+			logrus.Warn("certificate ou is empty - if you're using EIDAS certificates you need to configure issuer")
+		}
 		return ou + "/" + cn, nil
 	}