+++ pkg/model/specification_test.go
@@ -4,9 +4,8 @@ import (
 	"encoding/json"
 	"flag"
 	"io/ioutil"
-	"testing"
-
 	"path/filepath"
+	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"